requirements.txt
APScheduler
Flask
Flask_APScheduler
pandas
Pillow
Requests
SQLAlchemy
